console: Refactor device wizard
Prepare for device repository
