+++ pkg/webui/console/views/device-add/device-add.js
@@ -1,4 +1,4 @@
-// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2021 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,145 +13,56 @@
 // limitations under the License.
 
 import React from 'react'
-import { defineMessages } from 'react-intl'
+import { Switch, Route, Redirect } from 'react-router'
 import { Container, Col, Row } from 'react-grid-system'
-import { Switch, Route } from 'react-router'
+import { defineMessages } from 'react-intl'
 
 import PageTitle from '@ttn-lw/components/page-title'
-import Breadcrumb from '@ttn-lw/components/breadcrumbs/breadcrumb'
-import { withBreadcrumb } from '@ttn-lw/components/breadcrumbs/context'
+import Tabs from '@ttn-lw/components/tabs'
 
-import sharedMessages from '@ttn-lw/lib/shared-messages'
-import { getDeviceId } from '@ttn-lw/lib/selectors/id'
+import NotFoundRoute from '@ttn-lw/lib/components/not-found-route'
+
 import PropTypes from '@ttn-lw/lib/prop-types'
 
-import ConfigurationForm from './configuration-form'
-import DeviceWizard from './wizard'
+import Manual from './manual'
+import messages from './messages'
+
+import style from './device-add.styl'
 
 const m = defineMessages({
-  title: 'Add new end device',
+  title: 'Register end device',
 })
 
-const DeviceAdd = React.memo(props => {
-  const {
-    match,
-    location,
-    appId,
-    redirectToWizard,
-    redirectToEndDevice,
-    redirectToConfiguration,
-    redirectToNextLocation,
-    jsConfig,
-    nsConfig,
-    asConfig,
-    mayEditKeys,
-    createDevice,
-    prefixes,
-  } = props
+const DeviceAdd = props => {
+  const { match } = props
+  const { url } = match
 
-  const [configuration, setConfiguration] = React.useState(undefined)
-  const handleConfigurationSubmit = React.useCallback(
-    configuration => {
-      setConfiguration(configuration)
-
-      redirectToWizard()
-    },
-    [redirectToWizard],
+  const tabs = React.useMemo(
+    () => [
+      { title: messages.manualTabTitle, link: `${url}/manual`, name: 'manual' },
+    ],
+    [url],
   )
-  const hasConfiguration = typeof configuration !== 'undefined'
-
-  const rollbackProgress = React.useCallback(
-    nextLocation => {
-      redirectToNextLocation(nextLocation)
-    },
-    [redirectToNextLocation],
-  )
-
-  const handleCreateDevice = React.useCallback(
-    device => {
-      return createDevice(appId, device)
-    },
-    [appId, createDevice],
-  )
-
-  const handleCreateDeviceSuccess = React.useCallback(
-    device => {
-      const deviceId = getDeviceId(device)
-
-      redirectToEndDevice(appId, deviceId)
-    },
-    [appId, redirectToEndDevice],
-  )
-
-  React.useEffect(() => {
-    if (location.pathname.endsWith('/steps') && !hasConfiguration) {
-      redirectToConfiguration()
-    }
-  }, [hasConfiguration, location, redirectToConfiguration])
 
   return (
     <Container>
       <Row>
-        <Col lg={8} md={12}>
+        <Col>
           <PageTitle title={m.title} />
-          <Switch>
-            <Route exact path={`${match.url}`}>
-              {() => (
-                <ConfigurationForm
-                  asConfig={asConfig}
-                  jsConfig={jsConfig}
-                  nsConfig={nsConfig}
-                  onSubmit={handleConfigurationSubmit}
-                  initialValues={configuration}
-                  mayEditKeys={mayEditKeys}
-                />
-              )}
-            </Route>
-            <Route path={`${match.url}/steps`}>
-              {({ match }) =>
-                hasConfiguration ? (
-                  <DeviceWizard
-                    prefixes={prefixes}
-                    createDevice={handleCreateDevice}
-                    createDeviceSuccess={handleCreateDeviceSuccess}
-                    rollbackProgress={rollbackProgress}
-                    asConfig={asConfig}
-                    jsConfig={jsConfig}
-                    nsConfig={nsConfig}
-                    mayEditKeys={mayEditKeys}
-                    configuration={configuration}
-                    match={match}
-                  />
-                ) : null
-              }
-            </Route>
-          </Switch>
+          <Tabs className={style.tabs} narrow tabs={tabs} />
         </Col>
       </Row>
+      <Switch>
+        <Redirect exact from={url} to={`${url}/manual`} />
+        <Route path={`${match.url}/manual`} component={Manual} />
+        <NotFoundRoute />
+      </Switch>
     </Container>
   )
-})
+}
 
 DeviceAdd.propTypes = {
-  appId: PropTypes.string.isRequired,
-  asConfig: PropTypes.stackComponent.isRequired,
-  createDevice: PropTypes.func.isRequired,
-  jsConfig: PropTypes.stackComponent.isRequired,
-  location: PropTypes.location.isRequired,
   match: PropTypes.match.isRequired,
-  mayEditKeys: PropTypes.bool.isRequired,
-  nsConfig: PropTypes.stackComponent.isRequired,
-  prefixes: PropTypes.euiPrefixes,
-  redirectToConfiguration: PropTypes.func.isRequired,
-  redirectToEndDevice: PropTypes.func.isRequired,
-  redirectToNextLocation: PropTypes.func.isRequired,
-  redirectToWizard: PropTypes.func.isRequired,
 }
 
-DeviceAdd.defaultProps = {
-  prefixes: [],
-}
-
-export default withBreadcrumb('devices.add', props => (
-  <Breadcrumb path={`/applications/${props.appId}/devices/add`} content={sharedMessages.add} />
-))(DeviceAdd)
+export default DeviceAdd
