+++ pkg/webui/console/views/device-add/index.js
@@ -1,4 +1,4 @@
-// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
+// Copyright © 2021 The Things Network Foundation, The Things Industries B.V.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import DeviceAdd from './device-add'
 import connect from './connect'
+import DeviceAdd from './device-add'
 
 const ConnectedDeviceAdd = connect(DeviceAdd)
